// generatedTypes/ILSCheckingTransactionsDataSchema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ILSCheckingTransactionsDataSchema {
  type: {
    retrievalTransactionData: {
      bankNumber: {
        [k: string]: unknown;
      };
      retrievalMinDate: number;
      formattedRetrievalMinDate: string;
      retrievalMaxDate: number;
      formattedRetrievalMaxDate: string;
      retrievalStartDate: number;
      formattedRetrievalStartDate: string;
      retrievalEndDate: number;
      formattedRetrievalEndDate: string;
      balanceAmountDisplayIndication: {
        [k: string]: unknown;
      };
      branchNumber: number;
      accountNumber: number;
      joinPfm: {
        [k: string]: unknown;
      };
      eventCounter: number;
    };
    transactions: {
      eventDate: number;
      formattedEventDate: string;
      serialNumber: number;
      activityTypeCode: number;
      activityDescription: string;
      activityDescriptionIncludeValueDate: string | null;
      textCode: number;
      referenceNumber: number;
      referenceCatenatedNumber: number;
      valueDate: number;
      formattedValueDate: string;
      eventAmount: number;
      /**
       * Income (1) or expense (2)
       */
      eventActivityTypeCode: 1 | 2;
      currentBalance: number;
      internalLinkCode: number;
      originalEventCreateDate: number;
      formattedOriginalEventCreateDate: {
        [k: string]: unknown;
      };
      /**
       * Fixed (REGULAR) or still in process/temporary (TODAY)
       */
      transactionType: 'REGULAR' | 'TODAY';
      dataGroupCode: number;
      beneficiaryDetailsData?: {
        partyName?: string | null;
        messageHeadline?: string | null;
        partyHeadline?: string | null;
        messageDetail?: string | null;
        tableNumber?: number;
        recordNumber?: number;
      } | null;
      expandedEventDate: string;
      executingBranchNumber: number;
      eventId: 0 | 184;
      details: string | null;
      pfmDetails: {
        [k: string]: unknown;
      };
      differentDateIndication: string;
      rejectedDataEventPertainingIndication: string;
      tableNumber: number;
      recordNumber: number;
      contraBankNumber: number;
      contraBranchNumber: number;
      contraAccountNumber: number;
      contraAccountTypeCode: number;
      marketingOfferContext: {
        [k: string]: unknown;
      };
      commentExistenceSwitch: {
        [k: string]: unknown;
      };
      englishActionDesc: string | null;
      fieldDescDisplaySwitch: 0 | 1;
      offerActivityContext: string;
      comment: string | null;
    }[];
    comments: string[];
    message: string[];
    numItemsPerPage: number;
    pdfUrl: string;
  };
}